updated attributes

+++ route.sh
@@ -1,7 +1,22 @@
 #!/bin/bash
-# git workflow v7.0
-# 2022-2023 @arfazhxss
-# Modified for UVic Rocketry 
+
+# ┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
+# ┃                                 Git Workflow                                      ┃
+# ┃===================================================================================┃
+# ┃ Developed for this Github Repository: Tic Tac Toe Game.                           ┃
+# ┃ Tested on Mac, Fedora 36, and unix-based Distributions,                           ┃
+# ┃                                                                                   ┃
+# ┃ One-Keyboard Function(s): Pull/Push, Reset & Sync.                                ┃
+# ┃                                                                                   ┃
+# ┃ This is an open-source script. Feel free to use it at your own risk.              ┃
+# ┃                                                                                   ┃
+# ┃===================================================================================┃
+# ┃                    Suggestions/Others: <arfazhxss@duck.com>                       ┃
+# ┃===================================================================================┃
+# ┃                               Stable Script 8.1                                   ┃
+# ┃                                2024 @arfazhxss                                    ┃
+# ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
+
 
 breakStrSize=50
 breakStrIter=$(printf '_%.0s' $(seq 1 "$breakStrSize"))
@@ -24,8 +39,8 @@ function versionCheck() {
 function syncBranch() {
   echo -e "YES'ED\n${breakStrIter}"
   git stash
+  git pull --rebase --autostash --quiet
   git stash clear
-  git pull
   echo -e "${breakStrIter}\n\t\tYour Repository is synced\n\t\twith the latest commit :)\n${breakStrIter}"
 }
 
@@ -36,12 +51,14 @@ function pushChanges() {
   fi
 
   git add . && git add -u && \
-  git commit -m $"$CommitMessage"$'\nCommit by @arfazhxss on '"$(date +'%a %d %b %Y')" && \
-  git push --set-upstream origin "$branch"
-  rm -Rf .DS_Store/
-  echo -e "${breakStrIter}\n\t\tYour changes have been pushed\n\t\tto the repository :)\n${breakStrIter}"
+  git commit -m "$CommitMessage"$'\nCommit by @arfazhxss on '"$(date +'%a %d %b %Y')" && \
+  if git push --set-upstream origin "$branch" --quiet; then
+    rm -Rf .DS_Store/
+    echo -e "${breakStrIter}\n\t\tYour changes have been pushed\n\t\tto the repository :)\n${breakStrIter}"
+  else
+    echo -e "${breakStrIter}\n\t\tError: Failed to push changes to the repository\n${breakStrIter}"
+  fi
 }
-
 
 function gitWorkflow() {
   find . -name ".DS_Store" -type f -delete
